Switch from '%s' to '%o'
Making this change ensures that `util.inspect` is used to serialize the `extractedValue` and `result.captures[spec.as]` results.
This avoids an error `TypeError: Cannot convert object to primitive value` that occurs with some non-object entities that can be returned by the cheerioExtractor (i.e. DOM Elements which are not native objects).