fix(engine_http): Ensure config.http.pool works as expected
Refactor the logic that manages how socket connections are created
and managed. With this change, we no longer rely on request.js to
manage it for us, and instead either create a global agent to share
between all VUs, or create an agent for each VU depending on whether
config.http.pool is set.