core, lib: Expose ee, and allow IPC from worker processes
The hooks in the http engine were very useful, but they didn't provide outbound
communication to the parent process.  Users should be able to have the master
process handle things like file handles and external services such as rabbitmq.
By allowing `processor` functions to emit a 'custom' event to the master proc,
this can be bubbled up to the caller an allow extensibility such that the caller
can interface with other systems and tap into `.on('done')` to do things like
graceful shutdown, or even firing off other scenarios.