feat: remove "source" prefix from standard metric names
This change removes "core." and "engine." prefix from built-in metrics.

The motivation for this change is to follow the principle of least surprise.
Before this change, metric names seen by users in output on their terminal
when running Artillery did not correspond 1:1 to metric names inside
the generated JSON report.

It also makes the new ensure plugin more difficult to use, as metric names
shown in the terminal cannot be used in conditions - you have to know to
prepend "engine." or "core." where appropriate.

The original motivation to have these prefixes was to make grouping easier
in text output, i.e. all core metrics could be grouped together, but the
trade-off isn't worth it.

Another alternative is to display the prefixes, but it only makes the output
more difficult to read by adding what is essentially metric metadata to
Artillery's reports, and increases the amount of horizontal space needed.