fix(dockerfile): run npm link
Fixes #1394.

The currently built container image works fine if you want to
invoke artillery through a single docker command. But if you want to run
multiple artillery commands in the same container instance, or if you
want to use Artillery's container image as a base layer, then you ran
into issues such #1394.

The issue us that the executable in the path is named `run` instead of
`artillery`. To address this while maintain compatibility for any users
depending on the existing executable we can run an `npm install` which
will setup artillery as if it had been installed globally.