fix: rampTo distribution
Setting an arrival rate lower than the workers being user to run the test,
resulted in rampTo property being distributed across all workers instead
of the ones that had an actual arrivalRate.

As the default value for arrivalRate if not present was 1, it generated
a weird rampTo curve.

Changed the default to zero, in order to use all threads and keep the
expected ramp curve.