refactor: rework rampup logic
Use the biggest step that let us create the biggest arrival of the given
period.

Use a probability function that returns a value from 0 to 1 that
represents the probability of arrival at a given step.

In each posible arrival, roll a probability an compare it to
f(currentStep) to decide.

Also, this removes dead code