fix(bom): calculate common prefix based on filepath components
The current implementation of commonPrefix treats inputs as just
strings, whereas the intention is to return common *directory*
prefix for a bunch of filenames.

Current implementation breaks on inputs such as:

['/test/acme.js', '/test/acme.yml']

returning '/test/acme.' as the result.

This change fixes that issue. The result for the input above
will now be: '/test/'.

Ref: https://github.com/artilleryio/artillery/issues/2092