fix(aws-lambda): wait a little longer for the lambda when using a VPC
When using the aws-lambda platform for running tests and using a VPC, it often takes a little longer
for the function to become ready that it does when creating the lambda without a VPC. When the
lambda doesn't get created in time that leaves things in an inconsistent state, as the lambda is
also never cleaned up. To avoid changing the behaviour when not using a VPC, this change adds a
check before starting the wait for lambda flow. If the lambda is going to be connected to a VPC it
sets the wait timeout to 4 minutes, otherwise it sets it to the current value of 2 minutes.