feat: make values from --dotenv available to Azure workers
Currently the values set in an env file are only made available
to the Artillery process running in a worker Azure, instead of
being set at the container level.

This means that for example it's impossible to set the DEBUG
environment variable in the worker to help debug issues.

This change makes tests on Azure ACI work in the same way as
tests on AWS Fargate.