fix: remove files in Azure Blob Storage only once per test run
The current cleanup() code will attempt to remove files from Blob
Storage in every worker in a test. This can lead to the following
issue:

- User launches a test with a lot of workers (e.g. 100)
- Workers 1 to 50 launch and boostrap successfully
- Worker 51 encounters an error and exits early
- Workers 52 onwards won't be able to download
  test-related assets from Blob Storage anymore, and will fail

The main problem with this is that this makes debugging the actual
cause of a failure much harder, as container logs will be polluted
with errors related to inability to download test assets.