fix: add utility functions to get  module type
This PR will add more complete handling with regards to the CommonJS and ESM standards. This PR should solve #3203 and #2661.

We now look at the extension first, if it's .cjs or .mjs, great, that's easy to resolve. Otherwise, we look for the first package.json we come across, check it's module type, and then determine the project's standards.

The esbuild has been updated to accomodate this change as well.