fix: take distributed phases into account in TypeScript scripts
When the main script is a TypeScript file it gets transpiled in the main thread, and each worker thread then loads the transpiled module. This "resets" the results of divideWork() call, i.e. each worker runs with the original values of config.phases rather than the ones adjusted to distribute the work across all workers.

With this change, each worker receives a phase configuration for a subset of overall phase profile.

Fixes https://github.com/artilleryio/artillery/issues/3546