fix: make sure global.artillery.s3BucketRegion is set
This is needed because AWS SDK v3 requires that S3 clients have the right region set - otherwise a permanent redirect error is returned.

The region of the bucket will be determined by the region of the first Artillery test run in that account (i.e. the value of the --region flag used for the first time), which may be different on subsequent runs. So we have to look up the region of the bucket every time.