feat: add --no-assign-public-ip flag for Fargate
By default Artillery enables the assignPublicIp ENI option for worker tasks which assigns a public IP to each task. This is to make sure that tasks always have a route to the internet, which is required for pulling the container image.

This option is for use-cases where tasks need to run from a private subnet which has a NAT gateway attached to make sure that requests by VUs come from a known IP address.

More info: https://docs.aws.amazon.com/AmazonECS/latest/developerguide/fargate-task-networking.html